feat: 알림 권한 요청 기능 구현 및 네비게이션 구조 리팩터링
