+++ README.md
@@ -25,12 +25,12 @@
 
 # 5월 23일 구현할 기능 목록
 - [ ] 추천받은 영상을 학습하고 Crypto Manager를 변경하기 (예상 40분)
-- [ ] 알림 관련 기능 구현을 위한 학습(학습 1시간 반)
-  - 권한 관련 학습 (예상 30분)
+- [ ] 알림 관련 기능 구현을 위한 학습(예상 학습 1시간 반)
+  - 권한 관련 학습 (10분)
   - BroadcastReceiver 학습
   - AlarmManager 학습
   - SharedPreferences 학습
-- [ ] 알림을 위한 권한 받기 기능 구현 (예상 30분)
+- [x] 알림을 위한 권한 받기 기능 구현 (15분)
 - [ ] 알림 설정을 위한 UI 구현 (예상 1시간)
 - [x] 쿠폰 적용 규칙 구현 (1시간)
     - 1개만 적용 가능
